Move dataset/model actions to their own files; Add tests
